front: fixes sales catalog import
- extracts the sales catalog import component to its own file
